+++ hil-test/tests/rsa.rs
@@ -66,6 +66,7 @@ mod tests {
     }
 
     #[test]
+    #[timeout(5)]
     fn test_modular_exponentiation(mut ctx: Context<'static>) {
         const EXPECTED_OUTPUT: [u32; U512::LIMBS] = [
             1601059419, 3994655875, 2600857657, 1530060852, 64828275, 4221878473, 2751381085,
@@ -92,6 +93,7 @@ mod tests {
     }
 
     #[test]
+    #[timeout(5)]
     fn test_modular_multiplication(mut ctx: Context<'static>) {
         const EXPECTED_OUTPUT: [u32; U512::LIMBS] = [
             1868256644, 833470784, 4187374062, 2684021027, 191862388, 1279046003, 1929899870,
@@ -114,6 +116,7 @@ mod tests {
     }
 
     #[test]
+    #[timeout(5)]
     fn test_multiplication(mut ctx: Context<'static>) {
         const EXPECTED_OUTPUT: [u32; U1024::LIMBS] = [
             1264702968, 3552243420, 2602501218, 498422249, 2431753435, 2307424767, 349202767,